清理无用文件

+++ fe-xiaozhuye.js
@@ -9,8 +9,10 @@ const jsPath = './www/static/js/xiaozhuye';
 const chunkPath = './www/static/js/xiaozhuye/chunk';
 const indexHtmlPath = './view/home/index_index.html';
 const manifestPath = './www/tiny.appcache';
+
 // 清空js文件夹内所有文件
 // 删除chunk文件夹
+console.log('\033[91m 清理js文件夹... \033[0m' + '\n');
 if (fs.existsSync(chunkPath)) {
     let chunkFolder = fs.readdirSync(chunkPath);
     chunkFolder.forEach(fileName => {
@@ -26,9 +28,8 @@ if (fs.existsSync(jsPath)) {
     });
 }
 
-console.log('\033[91m 清理js文件夹; \033[0m' + '\n');
-
 // webpack打包
+console.log('\033[91m 开始打包js... \033[0m');
 compiler.run((err, stats) => {
     if (stats.hasErrors()) {
         return console.log(stats.toJson().errorDetail);
@@ -51,7 +52,7 @@ compiler.run((err, stats) => {
         // 更新文件
         fs.writeFile(indexHtmlPath, newScript, 'utf8', (err) => {
             if (err) throw err;
-            console.log('\033[92m' + indexHtmlPath + ' 更新成功; \033[0m' + '\n');
+            console.log('\033[92m ' + indexHtmlPath + ' 更新成功 \033[0m' + '\n');
         });
     });
 
@@ -59,40 +60,29 @@ compiler.run((err, stats) => {
     let newJsCache = '';
     _.each(chunksArr, (value) => {
         newJsCache += '\n' + '/static/js/' + value.files[0];
-        console.log('\033[92m 打包js完成; \033[0m' + value.files[0]);
+        console.log('\033[92m 打包js => \033[0m' + value.files[0] + '\033[92m 大小: ' + value.size / 1000 + 'KB\033[0m');
     });
 
     // 更新manifest配置
-    fs.readFile(manifestPath, 'utf8', (err, data) => {
-        if (err) throw err;
+    if (fs.existsSync(manifestPath)) {
+        console.log('\n\033[91m 检测到: ' + manifestPath + ' 开始清理旧文件... \033[0m');
+        fs.unlinkSync(manifestPath);
+        console.log('\033[91m 已删除: ' + manifestPath + ' \033[0m');
+    }
+    if (!fs.existsSync(manifestPath)) {
+        let verNum = '# ' + Date();
+        let verText = 'CACHE MANIFEST\n' + verNum;
+        let cssStatic = '\n/static/css/bootstrap.min.css\n' +
+            '/static/css/style.css\n' +
+            '/static/img/default.png\n' +
+            '/static/img/refresh.png\n' +
+            '/favicon.ico\n\n' +
+            'NETWORK:\n*';
+        let newFile = verText + newJsCache + cssStatic;
 
-        let jsExp = /\/static\/js\/\w*.*[\d\w]*\.js\n/g;
-        let verExp = /#\s\d+\-\d+\-\d+\sv\d\.\d\.\d\b/;
-        let verNumExp = /v\d\.\d\.\d\b/;
-
-        // 修改版本号
-        let currVer = data.match(verNumExp)[0];
-        let verArr = currVer.split('.');
-        let thirdArr = ++verArr[2];
-        let secondArr = verArr[1];
-        if (thirdArr >= 10) {
-            thirdArr = 0;
-            secondArr++;
-        }
-        let newVerNum = verArr[0] + '.' + secondArr + '.' + thirdArr;
-        // 版本号添加日期
-        let date = new Date();
-        let newVer = '# ' + date.getFullYear() + '-' + date.getMonth() + '-' + date.getDay() + ' ' + newVerNum;
-
-        // 删除原js文件
-        let newFile = data.replace(jsExp, '');
-        // 修改版本号,添加新js缓存
-        newFile = newFile.replace(verExp, newVer + newJsCache);
-
-        // 更新文件
         fs.writeFile(manifestPath, newFile, 'utf8', (err) => {
             if (err) throw err;
-            console.log('\033[92m' + manifestPath + ' 更新成功: \033[0m' + '\n' + newFile);
+            console.log('\033[92m 新建 ' + manifestPath + ' 成功 \033[0m' + '\n');
         });
-    });
+    }
 });
